Started Maze
Imported a crystal_mine asset which is all of the mine things.

Added a script to the camera so you can walk around a bit.

Built a wee bit of the maze, added some glowy things.

Untick of the directional light to play through, it makes it way more creepy.